feat: プランカード・スポットカードにスポットピン表示機能を追加 (#278)
- プランカードのスポット行クリックで地図に単一スポットをピン表示
- スポットカードにピンボタン追加
- 「地図で見る」を「ルートを見る」に変更（差別化）
- community_preview.jsをStimulusコントローラに統合
- single_spot_preview_controller.js新規作成
