+++ app/views/plans/_community_tab.html.erb
@@ -141,7 +141,9 @@
                 spot = ps.spot
                 {
                   name: spot.name,
-                  address: spot.address
+                  address: spot.address,
+                  lat: spot.lat,
+                  lng: spot.lng
                 }
               end
               # 全スポットからユニークなジャンル名を収集
